refactor + fix tests
