+++ crates/learner/src/retriever/mod.rs
@@ -8,7 +8,7 @@
 //!
 //! The retriever system consists of several key components:
 //!
-//! - [`Retriever`]: Main entry point for paper retrieval operations
+//! - [`Retrievers`]: Main entry point for paper retrieval operations
 //! - [`RetrieverConfig`]: Configuration for specific paper sources
 //! - [`ResponseFormat`]: Format-specific parsing logic (XML/JSON)
 //! - [`ResponseProcessor`]: Trait for processing API responses
@@ -35,12 +35,15 @@
 //! Configure and use a retriever:
 //!
 //! ```no_run
-//! use learner::retriever::{Retriever, RetrieverConfig};
+//! use learner::{
+//!   prelude::*,
+//!   retriever::{RetrieverConfig, Retrievers},
+//! };
 //!
 //! # async fn example() -> Result<(), Box<dyn std::error::Error>> {
 //! // Create a new retriever
 //! let retriever =
-//!   Retriever::new().with_config_file("config/arxiv.toml")?.with_config_file("config/doi.toml")?;
+//!   Retrievers::new().with_config_file("config/arxiv.toml")?.with_config_file("config/doi.toml")?;
 //!
 //! // Retrieve a paper (automatically detects source)
 //! let paper = retriever.get_paper("10.1145/1327452.1327492").await?;
@@ -52,10 +55,11 @@
 //! Load multiple configurations:
 //!
 //! ```no_run
-//! # use learner::retriever::Retriever;
+//! # use learner::retriever::Retrievers;
+//! # use learner::prelude::*;
 //! # async fn example() -> Result<(), Box<dyn std::error::Error>> {
 //! // Load all TOML configs from a directory
-//! let retriever = Retriever::new().with_config_dir("config/")?;
+//! let retriever = Retrievers::new().with_config_dir("config/")?;
 //!
 //! // Retriever will automatically match source based on input format
 //! let arxiv_paper = retriever.get_paper("2301.07041").await?;
@@ -84,9 +88,10 @@ pub use response::*;
 /// # Examples
 ///
 /// ```no_run
-/// # use learner::retriever::Retriever;
+/// # use learner::retriever::Retrievers;
+/// # use learner::prelude::*;
 /// # async fn example() -> Result<(), Box<dyn std::error::Error>> {
-/// let retriever = Retriever::new().with_config_dir("config/")?;
+/// let retriever = Retrievers::new().with_config_dir("config/")?;
 ///
 /// // Retrieve papers from different sources
 /// let paper1 = retriever.get_paper("2301.07041").await?; // arXiv
@@ -96,20 +101,18 @@ pub use response::*;
 /// # }
 /// ```
 #[derive(Default, Debug, Clone)]
-pub struct Retriever {
+pub struct Retrievers {
   /// The collection of configurations used for this [`Retriever`].
-  configs: HashMap<String, RetrieverConfig>,
+  configs: BTreeMap<String, RetrieverConfig>,
 }
 
-impl Configurable for Retriever {
+impl Configurable for Retrievers {
   type Config = RetrieverConfig;
 
-  fn insert(&mut self, config_name: String, config: Self::Config) {
-    self.configs.insert(config_name, config);
-  }
+  fn as_map(&mut self) -> &mut BTreeMap<String, Self::Config> { &mut self.configs }
 }
 
-impl Retriever {
+impl Retrievers {
   /// Checks whether the retreivers map is empty.
   ///
   /// This is useful for handling the case where no retreivers are specified and
@@ -118,11 +121,11 @@ impl Retriever {
   /// # Examples
   ///
   /// ```no_run
-  /// # use learner::retriever::Retriever;
+  /// # use learner::retriever::Retrievers;
   /// # use learner::error::LearnerError;
   ///
   /// # fn check_is_empty() -> Result<(), LearnerError> {
-  /// let retriever = Retriever::new();
+  /// let retriever = Retrievers::new();
   ///
   /// if retriever.is_empty() {
   ///   return Err(LearnerError::Config("No retriever configured.".to_string()));
@@ -133,15 +136,15 @@ impl Retriever {
   pub fn is_empty(&self) -> bool { self.configs.is_empty() }
 }
 
-impl Retriever {
+impl Retrievers {
   /// Creates a new empty retriever with no configurations.
   ///
   /// # Examples
   ///
   /// ```no_run
-  /// use learner::retriever::Retriever;
+  /// use learner::retriever::Retrievers;
   ///
-  /// let retriever = Retriever::new();
+  /// let retriever = Retrievers::new();
   /// ```
   pub fn new() -> Self { Self::default() }
 
@@ -170,9 +173,10 @@ impl Retriever {
   /// # Examples
   ///
   /// ```no_run
-  /// # use learner::retriever::Retriever;
+  /// # use learner::retriever::Retrievers;
+  /// # use learner::prelude::*;
   /// # async fn example() -> Result<(), Box<dyn std::error::Error>> {
-  /// let retriever = Retriever::new().with_config_dir("config/")?;
+  /// let retriever = Retrievers::new().with_config_dir("config/")?;
   ///
   /// // Retrieve from different sources
   /// let paper1 = retriever.get_paper("2301.07041").await?;
@@ -226,9 +230,10 @@ impl Retriever {
   /// # Examples
   ///
   /// ```
-  /// # use learner::retriever::Retriever;
+  /// # use learner::retriever::Retrievers;
+  /// # use learner::prelude::*;
   /// # async fn example() -> Result<(), Box<dyn std::error::Error>> {
-  /// let retriever = Retriever::new().with_config_dir("config/")?;
+  /// let retriever = Retrievers::new().with_config_dir("config/")?;
   ///
   /// // Sanitize an arXiv URL
   /// let (source, id) = retriever.sanitize_identifier("https://arxiv.org/abs/2301.07041")?;
@@ -320,5 +325,191 @@ fn apply_transform(value: &str, transform: &Transform) -> Result<String> {
         .map(|dt| dt.format(to_format).to_string()),
     Transform::Url { base, suffix } =>
       Ok(format!("{}{}", base.replace("{value}", value), suffix.as_deref().unwrap_or(""))),
+  }
+}
+
+#[cfg(test)]
+mod tests {
+  use super::*;
+
+  #[test]
+  fn validate_arxiv_config() {
+    let config_str = include_str!("../../config/retrievers/arxiv.toml");
+
+    let retriever: RetrieverConfig = toml::from_str(config_str).expect("Failed to parse config");
+
+    // Verify basic fields
+    assert_eq!(retriever.name, "arxiv");
+    assert_eq!(retriever.base_url, "http://export.arxiv.org");
+    assert_eq!(retriever.source, "arxiv");
+
+    // Test pattern matching
+    assert!(retriever.pattern.is_match("2301.07041"));
+    assert!(retriever.pattern.is_match("math.AG/0601001"));
+    assert!(retriever.pattern.is_match("https://arxiv.org/abs/2301.07041"));
+    assert!(retriever.pattern.is_match("https://arxiv.org/pdf/2301.07041"));
+    assert!(retriever.pattern.is_match("https://arxiv.org/abs/math.AG/0601001"));
+    assert!(retriever.pattern.is_match("https://arxiv.org/abs/math/0404443"));
+
+    // Test identifier extraction
+    assert_eq!(retriever.extract_identifier("2301.07041").unwrap(), "2301.07041");
+    assert_eq!(
+      retriever.extract_identifier("https://arxiv.org/abs/2301.07041").unwrap(),
+      "2301.07041"
+    );
+    assert_eq!(retriever.extract_identifier("math.AG/0601001").unwrap(), "math.AG/0601001");
+
+    // Verify response format
+
+    if let ResponseFormat::Xml(config) = &retriever.response_format {
+      assert!(config.strip_namespaces);
+
+      // Verify field mappings
+      let field_maps = &config.field_maps;
+      assert!(field_maps.contains_key("title"));
+      assert!(field_maps.contains_key("abstract"));
+      assert!(field_maps.contains_key("authors"));
+      assert!(field_maps.contains_key("publication_date"));
+      assert!(field_maps.contains_key("pdf_url"));
+
+      // Verify PDF transform
+      if let Some(map) = field_maps.get("pdf_url") {
+        match &map.transform {
+          Some(Transform::Replace { pattern, replacement }) => {
+            assert_eq!(pattern, "/abs/");
+            assert_eq!(replacement, "/pdf/");
+          },
+          _ => panic!("Expected Replace transform for pdf_url"),
+        }
+      } else {
+        panic!("Missing pdf_url field map");
+      }
+    } else {
+      panic!("Expected an XML configuration, but did not get one.")
+    }
+
+    // Verify headers
+    assert_eq!(retriever.headers.get("Accept").unwrap(), "application/xml");
+  }
+
+  #[test]
+  fn test_doi_config_deserialization() {
+    let config_str = include_str!("../../config/retrievers/doi.toml");
+
+    let retriever: RetrieverConfig = toml::from_str(config_str).expect("Failed to parse config");
+
+    // Verify basic fields
+    assert_eq!(retriever.name, "doi");
+    assert_eq!(retriever.base_url, "https://api.crossref.org/works");
+    assert_eq!(retriever.source, "doi");
+
+    // Test pattern matching
+    let test_cases = [
+      ("10.1145/1327452.1327492", true),
+      ("https://doi.org/10.1145/1327452.1327492", true),
+      ("invalid-doi", false),
+      ("https://wrong.url/10.1145/1327452.1327492", false),
+    ];
+
+    for (input, expected) in test_cases {
+      assert_eq!(
+        retriever.pattern.is_match(input),
+        expected,
+        "Pattern match failed for input: {}",
+        input
+      );
+    }
+
+    // Test identifier extraction
+    assert_eq!(
+      retriever.extract_identifier("10.1145/1327452.1327492").unwrap(),
+      "10.1145/1327452.1327492"
+    );
+    assert_eq!(
+      retriever.extract_identifier("https://doi.org/10.1145/1327452.1327492").unwrap(),
+      "10.1145/1327452.1327492"
+    );
+
+    // Verify response format
+    match &retriever.response_format {
+      ResponseFormat::Json(config) => {
+        // Verify field mappings
+        let field_maps = &config.field_maps;
+        assert!(field_maps.contains_key("title"));
+        assert!(field_maps.contains_key("abstract"));
+        assert!(field_maps.contains_key("authors"));
+        assert!(field_maps.contains_key("publication_date"));
+        assert!(field_maps.contains_key("pdf_url"));
+        assert!(field_maps.contains_key("doi"));
+      },
+      _ => panic!("Expected JSON response format"),
+    }
+  }
+
+  #[test]
+  fn test_iacr_config_deserialization() {
+    let config_str = include_str!("../../config/retrievers/iacr.toml");
+
+    let retriever: RetrieverConfig = toml::from_str(config_str).expect("Failed to parse config");
+
+    // Verify basic fields
+    assert_eq!(retriever.name, "iacr");
+    assert_eq!(retriever.base_url, "https://eprint.iacr.org");
+    assert_eq!(retriever.source, "iacr");
+
+    // Test pattern matching
+    let test_cases = [
+      ("2016/260", true),
+      ("2023/123", true),
+      ("https://eprint.iacr.org/2016/260", true),
+      ("https://eprint.iacr.org/2016/260.pdf", true),
+      ("invalid/format", false),
+      ("https://wrong.url/2016/260", false),
+    ];
+
+    for (input, expected) in test_cases {
+      assert_eq!(
+        retriever.pattern.is_match(input),
+        expected,
+        "Pattern match failed for input: {}",
+        input
+      );
+    }
+
+    // Test identifier extraction
+    assert_eq!(retriever.extract_identifier("2016/260").unwrap(), "2016/260");
+    assert_eq!(
+      retriever.extract_identifier("https://eprint.iacr.org/2016/260").unwrap(),
+      "2016/260"
+    );
+    assert_eq!(
+      retriever.extract_identifier("https://eprint.iacr.org/2016/260.pdf").unwrap(),
+      "2016/260"
+    );
+
+    // Verify response format
+    if let ResponseFormat::Xml(config) = &retriever.response_format {
+      assert!(config.strip_namespaces);
+
+      // Verify field mappings
+      let field_maps = &config.field_maps;
+      assert!(field_maps.contains_key("title"));
+      assert!(field_maps.contains_key("abstract"));
+      assert!(field_maps.contains_key("authors"));
+      assert!(field_maps.contains_key("publication_date"));
+      assert!(field_maps.contains_key("pdf_url"));
+
+      // Verify OAI-PMH paths
+      if let Some(map) = field_maps.get("title") {
+        assert!(map.path.contains(&"OAI-PMH/GetRecord/record/metadata/dc/title".to_string()));
+      } else {
+        panic!("Missing title field map");
+      }
+    } else {
+      panic!("Expected an XML configuration, but did not get one.")
+    }
+
+    // Verify headers
+    assert_eq!(retriever.headers.get("Accept").unwrap(), "application/xml");
   }
 }
