+++ crates/learner/tests/workflows/mod.rs
@@ -2,6 +2,5 @@ use learner::retriever::RetrieverConfig;
 
 use super::*;
 
-mod build_retriever;
 mod database_operations;
 mod paper_retrieval;
